% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndices.R
\docType{package}
\name{EmpiricTherapyIndices-package}
\alias{EmpiricTherapyIndices}
\alias{EmpiricTherapyIndices-package}
\title{Calculate empiric therapy indices for device-associated infections.}
\description{
The method is described in Hughes et al. 2016. Empiric therapy indices: novel measures of the impact of antibiotic resistance on populations.
}
\details{
The only functions you're likely to need from \pkg{EmpiricTherapyIndices} are
\code{\link{getExampleAntibiogramDAI}} and \code{\link{EmpiricTherapyIndicesDAI}}.
}

