% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndicesFns.R
\name{EmpiricTherapyIndicesDAI}
\alias{EmpiricTherapyIndicesDAI}
\title{Calculate empiric therapy indices for device-associated infections.}
\usage{
EmpiricTherapyIndicesDAI(Antibiograms, Basket = getBasketDAI(),
  Drugs = getDrugsDAI(), wide = T)
}
\arguments{
\item{Antibiograms}{An antibiogram. getExampleAntibiograms() for example. Numbers are percent resistance (or non-susceptibility). One site is included in the example, but the algorithm can handle multiple sites.}

\item{Basket}{The basket of infections. getBasketDAI() for example.}

\item{Drugs}{The set of available drugs. getDrugsDAI() for example.}
}
\value{
Empiric options index (EOI) and empiric coverage index (ECI) for each site.
}
\description{
Calculate empiric therapy indices for device-associated infections.
}
\examples{
EmpiricTherapyIndicesDAI(getExampleAntibiogramDAI())
}

