% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndicesFns.R
\name{getBasketDAI}
\alias{getBasketDAI}
\title{The basket of device-associated infections.}
\usage{
getBasketDAI()
}
\value{
The basket of device-associate drugs.
}
\description{
The basket of device-associated infections.
}
\examples{
Basket=getBasketDAI()
}

