% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndicesFns.R
\name{getDrugsDAI}
\alias{getDrugsDAI}
\title{Available drugs for device associated infections.}
\usage{
getDrugsDAI()
}
\value{
The set of available drugs.
}
\description{
Available drugs for device associated infections.
}
\examples{
Drugs=getDrugsDAI()
}

