% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndicesFns.R
\name{getExampleAntibiogramDAI}
\alias{getExampleAntibiogramDAI}
\title{An example antibiogram.}
\usage{
getExampleAntibiogramDAI()
}
\value{
An example antibiogram. Numbers are percent resistant (or non-susceptible).
}
\description{
An example antibiogram.
}
\examples{
#Revise the antibiogram and recalculate. PDR Acinetobacter example
pdrAcinetoAbg = getExampleAntibiogramDAI()
pdrAcinetoAbg$A.baumannii[!is.na(pdrAcinetoAbg$A.baumannii)]=100
EmpiricTherapyIndicesDAI(pdrAcinetoAbg)

#Save the example antibiogram as a delimited text file that can be edited in Excel
write.table(getExampleAntibiogramDAI(),file="myAntibiogram.txt",sep = "\\t",row.names=F)

myRevisedAbg = read.table(file="myAntibiogram.txt",header=T)
EmpiricTherapyIndicesDAI(myRevisedAbg)

#Calculate the indices without reserve drugs:
ReserveDrugs = c("amika","cefta","dapto","eryth","genta","linez","merop","tobra")
abgWithoutReserves = subset(getExampleAntibiogramDAI(),!is.element(drug_d,ReserveDrugs))
EmpiricTherapyIndicesDAI(abgWithoutReserves)
}

