% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndicesFns.R
\name{EmpiricTherapyIndicesDAI}
\alias{EmpiricTherapyIndicesDAI}
\title{Calculate empiric therapy indices for device-associated infections.}
\usage{
EmpiricTherapyIndicesDAI(Antibiograms, Basket = getBasketDAI(),
  Drugs = getDrugsDAI(), mixingPars = list(alpha = 0), wide = T)
}
\arguments{
\item{Antibiograms}{An antibiogram. getExampleAntibiograms() for example. Numbers are percent resistance (or non-susceptibility). One site is included in the example, but the algorithm can handle multiple sites.}

\item{Basket}{The basket of infections. getBasketDAI() for example.}

\item{Drugs}{The set of available drugs. getDrugsDAI() for example.}

\item{mixingPars}{A list of parameters for mixingFunction(), determining the impact of drug heterogeneity on the evolution of resistance. In the default case list(alpha=0) using multiple drugs does not slow the development of resistance. list(alpha=0.1) means that using 2 drugs instead of 1 slows the developement of resistance by 10\%.}
}
\value{
Empiric options index (EOI) and empiric coverage index (ECI) for each site.
}
\description{
Calculate empiric therapy indices for device-associated infections.
}
\examples{
#Example calculation:
EmpiricTherapyIndicesDAI(getExampleAntibiogramDAI())

#Revise the antibiogram and recalculate. PDR Acinetobacter example
pdrAcinetoAbg = getExampleAntibiogramDAI()
pdrAcinetoAbg$A.baumannii[!is.na(pdrAcinetoAbg$A.baumannii)]=100
EmpiricTherapyIndicesDAI(pdrAcinetoAbg)

#Save the example antibiogram as a delimited text file that can be edited in Excel
write.table(getExampleAntibiogramDAI(),file="myAntibiogram.txt",sep = "\\t",row.names=F)

myRevisedAbg = read.table(file="myAntibiogram.txt",header=T)
EmpiricTherapyIndicesDAI(myRevisedAbg)

#Calculate the indices without reserve drugs
ReserveDrugs = c("amika","cefta","dapto","eryth","genta","linez","merop","tobra")
abgWithoutReserves = subset(getExampleAntibiogramDAI(),!is.element(drug_d,ReserveDrugs))
EmpiricTherapyIndicesDAI(abgWithoutReserves)
}

