% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricTherapyIndicesFns.R
\name{getExampleAntibiogramDAI}
\alias{getExampleAntibiogramDAI}
\title{An example antibiogram.}
\usage{
getExampleAntibiogramDAI()
}
\value{
An example antibiogram. Numbers are percent resistant (or non-susceptible).
}
\description{
An example antibiogram.
}
\examples{
Antibiograms=getExampleAntibiogramDAI()
indices = EmpiricTherapyIndicesDAI(eAntibiograms)
}

